refactor:폴더 구조 변경 -#7
