+++ src/main/java/com/mju_lion/letter/dto/request/auth/LoginDto.java
@@ -1,4 +1,4 @@
-package com.mju_lion.letter.dto;
+package com.mju_lion.letter.dto.request.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
